requirements.txt
mailersend
python-dotenv
requests